refactor: modularize player actions
